safari-534.53-branch/LayoutTests/sputnik/Conformance/11_Expressions/11.13_Assignment_Operators/11.13.1_Simple_Assignment/S11.13.1_A2.1_T3-expected.txt

S11.13.1_A2.1_T3

PASS ReferenceError: Left side of assignment is not a reference.

TEST COMPLETE
